Add spendingInfoDb json serializer test, does not pass
